菜品相关接口：批量删除完成

+++ sky-server/src/main/java/pers/lwb/mapper/SetmealMapper.java
@@ -5,6 +5,6 @@ import org.apache.ibatis.annotations.Mapper;
 @Mapper
 public interface SetmealMapper {
 
-    Integer countByCategoryId(Long id);
+    Long countByCategoryId(Long id);
 
 }
